devproject/requirements.txt
Django==1.6.5
backports.ssl-match-hostname==3.4.0.2
django-admin-bootstrapped==1.6.5
django-crispy-forms==1.4.0
django-tornado==2.0.0a
tornado==3.2.2
usethis-django-bootstrap==3.2.0
wsgiref==0.1.2
